+++ src/db/query/mkvdrama.ts
@@ -1,3 +1,4 @@
+import { eq } from "drizzle-orm";
 import { handleError } from "../../utils/error.js";
 import { Logger } from "../../utils/logger.js";
 import { db } from "../drizzle.js";
@@ -12,4 +13,15 @@ export async function upsertMkvdrama(mkvdramas: EMkvdramaInsert[]) {
   } catch (e) {
     handleError(e, logger, `Failed to upsert mkvdrama`);
   }
+}
+
+export function getMkvdrama(providerContentId: string) {
+  if (!db) return;
+  const row = db.query.mkvdrama.findMany({
+    where: eq(mkvdrama.providerContentId, providerContentId),
+    with: {
+      ouo: true,
+    },
+  });
+  return row;
 }