+++ src/service/provider/provider-service.ts
@@ -1,13 +1,20 @@
 import { ContentType } from "@stremio-addon/sdk";
 import {
+  getContentByTmdb,
   getContentJoinProviderById,
   getCountProviderContent,
-  getProviderContent,
   getProviderContentsById as getMatchContentProvider,
+  getProviderContent,
+  getProviderContentById,
+  upsertContent,
+  upsertProviderContent,
 } from "../../db/queries.js";
+import { Prefix } from "../../lib/manifest.js";
 import { ContentDetail, ContentId } from "../../source/meta.js";
 import { Provider } from "../../source/provider.js";
-import { Prefix } from "../../lib/manifest.js";
+import { extractTitle } from "../../utils/format.js";
+import { tmdb } from "../../source/tmdb.js";
+import { TTL_MS } from "../../utils/cache.js";
 
 class ProviderService {
   static async getProviderContent(id: string) {
@@ -77,18 +84,24 @@ class ProviderService {
     return null;
   }
 
-  static async getDbContent(contentType: ContentType, contentId: ContentId) {
+  static async getDbContent(
+    type: ContentType,
+    contentId: ContentId,
+    season: number,
+  ) {
     let dbContent: ContentDetail | null = null;
-    const { imdbId, tmdbId, tvdbId, kisskhId, onetouchtvId } = contentId;
+    const { imdbId, tmdbId, tvdbId, kisskhId, onetouchtvId, mkvdramaId } =
+      contentId;
     const id: string | undefined =
       imdbId ??
       tmdbId?.toString() ??
       tvdbId?.toString() ??
       kisskhId ??
-      onetouchtvId;
+      onetouchtvId ??
+      mkvdramaId;
     if (!id) return null;
     const dbContentAndProvider = await ProviderService.getContentAndProvider(
-      contentType,
+      type,
       contentId,
     );
     if (!dbContentAndProvider) return null;
@@ -97,30 +110,114 @@ class ProviderService {
       id: id,
       title: dbContentAndProvider.title,
       year: dbContentAndProvider.year,
-      type: contentType,
+      type: type,
     };
-    if (contentId.imdbId) dbContent.imdbId = contentId.imdbId;
-    if (contentId.tmdbId) dbContent.tmdbId = contentId.tmdbId;
-    if (contentId.tvdbId) dbContent.tvdbId = contentId.tvdbId;
-    if (contentId.kisskhId) dbContent.kisskhId = contentId.kisskhId;
-    if (contentId.onetouchtvId) dbContent.onetouchtvId = contentId.onetouchtvId;
-    if (contentId.mkvdramaId) dbContent.mkvdramaId = contentId.mkvdramaId;
+    if (imdbId) dbContent.imdbId = imdbId;
+    if (tmdbId) dbContent.tmdbId = tmdbId;
+    if (tvdbId) dbContent.tvdbId = tvdbId;
+    if (dbContentAndProvider.imdbId)
+      dbContent.imdbId = dbContentAndProvider.imdbId;
+    if (dbContentAndProvider.tmdbId)
+      dbContent.tmdbId = parseInt(dbContentAndProvider.tmdbId);
+    if (dbContentAndProvider.tvdbId)
+      dbContent.tvdbId = parseInt(dbContentAndProvider.tvdbId);
+
+    // Set season from the correct provider
+    if (kisskhId) {
+      dbContent.kisskhId = kisskhId;
+      const provider = dbProviderContent.find((providerContent) => {
+        return providerContent.externalId === kisskhId;
+      });
+      if (provider) {
+        const extractedTitle = extractTitle(provider.title);
+        season = extractedTitle.season ?? 1;
+      }
+    }
+    if (onetouchtvId) {
+      dbContent.onetouchtvId = onetouchtvId;
+      const provider = dbProviderContent.find((providerContent) => {
+        return providerContent.externalId === onetouchtvId;
+      });
+      if (provider) {
+        const extractedTitle = extractTitle(provider.title);
+        season = extractedTitle.season ?? 1;
+      }
+    }
+    if (mkvdramaId) {
+      dbContent.mkvdramaId = mkvdramaId;
+      const provider = dbProviderContent.find((providerContent) => {
+        return providerContent.externalId === mkvdramaId;
+      });
+      if (provider) {
+        const extractedTitle = extractTitle(provider.title);
+        season = extractedTitle.season ?? 1;
+      }
+    }
     dbProviderContent.forEach((providerContent) => {
+      const extractedTitle = extractTitle(providerContent.title);
+      const extractedSeason = extractedTitle.season ?? 1;
       switch (providerContent.provider) {
         case Provider.KISSKH:
-          dbContent.kisskhId = providerContent.externalId;
+          if (!kisskhId && extractedSeason === season) {
+            dbContent.kisskhId = providerContent.externalId;
+            dbContent.title = providerContent.title;
+          }
           break;
         case Provider.ONETOUCHTV:
-          dbContent.onetouchtvId = providerContent.externalId;
+          if (!onetouchtvId && extractedSeason === season) {
+            dbContent.onetouchtvId = providerContent.externalId;
+            dbContent.title = providerContent.title;
+          }
           break;
         case Provider.MKVDRAMA:
-          dbContent.mkvdramaId = providerContent.externalId;
+          if (!mkvdramaId && extractedSeason === season) {
+            dbContent.mkvdramaId = providerContent.externalId;
+            dbContent.title = providerContent.title;
+          }
           break;
         default:
           break;
       }
     });
     return dbContent;
+  }
+
+  static async syncContentAndProvider(
+    content: ContentDetail,
+    id: string,
+    provider: Provider,
+  ) {
+    const title = content.title;
+    const tmdbDetail = await tmdb.searchDetail(content.title, content.type);
+    const image = content.thumbnail ?? tmdbDetail?.thumbnail;
+    const providerContentId = `${Prefix.MKVDRAMA}:${id}`;
+    if (tmdbDetail) {
+      const oldContent = await getContentByTmdb(tmdbDetail.id, content.type);
+      if (oldContent) {
+        const contentId = oldContent.id;
+        upsertContent(contentId, tmdbDetail, TTL_MS.content);
+        const providerContent = await getProviderContentById(providerContentId);
+        if (providerContent) {
+          await upsertProviderContent({
+            ...providerContent,
+            contentId: contentId,
+            title: title,
+          });
+        } else {
+          await upsertProviderContent({
+            id: providerContentId,
+            contentId: contentId,
+            title: title,
+            ttl: null,
+            provider: provider,
+            externalId: id.toString(),
+            image: image,
+            year: content.year,
+            type: content.type,
+          });
+        }
+      }
+    }
   }
 }